Organize $lib folder
