+++ src/lib/components/LoadingBanner.svelte
