+++ src/lib/stores.ts
@@ -1,13 +1,13 @@
 import type { Writable } from 'svelte/store';
 import { writable } from 'svelte/store';
 
-import type { Attendance } from '$lib/Attendance';
-import type { Gradebook } from '$lib/Gradebook';
-import type { ReportCardListEntity } from '$lib/ReportCardListEntity';
-import type { StudentInfo } from '$lib/StudentInfo';
 import type { StudentAccount } from '$lib/synergy';
-import type { DocumentsList } from './DocumentsList';
-import type { Message } from './Message';
+import type { Attendance } from '$lib/types/Attendance';
+import type { Gradebook } from '$lib/types/Gradebook';
+import type { ReportCardListEntity } from '$lib/types/ReportCardListEntity';
+import type { StudentInfo } from '$lib/types/StudentInfo';
+import type { DocumentsList } from '$lib/types/DocumentsList';
+import type { Message } from './types/Message';
 
 export const studentAccount: Writable<StudentAccount | undefined> = writable();
