+++ src/routes/(authed)/+layout.svelte
@@ -2,7 +2,7 @@
 	import { browser } from '$app/environment';
 	import { goto } from '$app/navigation';
 	import { navigating } from '$app/stores';
-	import AppSidebar from '$lib/AppSidebar.svelte';
+	import AppSidebar from '$lib/components/AppSidebar.svelte';
 	import { studentAccount } from '$lib/stores';
 	import { StudentAccount } from '$lib/synergy';
 	import { Drawer, Navbar, NavBrand, NavHamburger } from 'flowbite-svelte';
