+++ src/routes/(authed)/grades/+layout.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { browser } from '$app/environment';
-	import LoadingBanner from '$lib/LoadingBanner.svelte';
 	import { loadGradebook } from '$lib/cache';
+	import LoadingBanner from '$lib/components/LoadingBanner.svelte';
 	import { gradebook, gradebookLoaded } from '$lib/stores';
 
 	if (!$gradebook && browser) loadGradebook();
